Update and rename test_imguibindings.py to test_loadingplugins.py
